Rename helpers.ts to lib/utils.ts
