+++ src/components/dashboard.tsx
@@ -3,7 +3,7 @@ import { Heading } from "./ui/heading";
 import { HTMLAttributes } from "react";
 import { Section } from "./ui/section";
 import { VariantProps, cva } from "class-variance-authority";
-import { cn } from "@/utils/helpers";
+import { cn } from "../../lib/utils";
 
 interface DashboardRootLayoutProps {
   children: React.ReactNode;
