+++ src/components/ui/pretitle.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/utils/helpers";
+import { cn } from "../../../lib/utils";
 import { VariantProps, cva } from "class-variance-authority";
 import React, { FC } from "react";
 import { HeadingProps } from "./heading";
